New functional test for sharing, move session creation to base
