+++ functional_tests/lists/test_my_lists.py
@@ -1,28 +1,10 @@
-from django.conf import settings
 from django.contrib.auth import BACKEND_SESSION_KEY, SESSION_KEY, get_user_model
 from .lists_base import FunctionalTest
-from functional_tests.server_tools import create_session_on_server
-from functional_tests.management.commands.create_session import create_pre_authenticated_session
-from django.contrib.sessions.backends.db import SessionStore
 
 User = get_user_model()
 
 
 class MyListsTest(FunctionalTest):
-
-    def create_pre_authenticated_session(self, email):
-        if self.against_staging:
-            session_key = create_session_on_server(self.server_host, email)
-        else:
-            session_key = create_pre_authenticated_session(email)
-        ## to set a cookie we need to first visit the domain.
-        ## 404 pages load the quickest!
-        self.browser.get(self.server_url + "/404_no_such_url/")
-        self.browser.add_cookie(dict(
-            name=settings.SESSION_COOKIE_NAME,
-            value=session_key,
-            path='/',
-        ))
 
     def test_logged_in_users_lists_are_saved_as_my_lists(self):
         # Edith is a logged-in user
